+++ CHANGES.rst
@@ -1,6 +1,10 @@
 Changes
 =======
 
+- Fix debug mode with disk caching off where temporary caches were created in
+  the wrong place (i.e. alongside the file rather than in a temporary
+  directory).  [jinty]
+
 - Fixed issue with eager loading and template initialization [aikom].
 
 - Fix mishandling of XML declaration. (LP#360296)